requirements.txt
beautifulsoup4==4.11.1
python-slugify==7.0.0
Pillow==9.3.0
selenium==4.6.1
pandas==1.5.2
requests==2.28.1
webdriver-manager==3.8.5